% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare two dataframes for variables that match or don't match}
\usage{
compare(new, old, ...)
}
\arguments{
\item{new}{The "new" dataframe.}

\item{old}{The "old" dataframe to compare the new dataframe to.}

\item{...}{The ID column(s) to be that indicate a unique record to compare between dataframes.}
}
\description{
This function compares all of the variables in two similar dataframes to determine if the values in the data have changed.
This is helpful for identifying changes when a dataset is updated, similar to SAS PROC COMPARE.
}
\examples{
compare(new=new_df, old_df, id)
}
\keyword{Compare}
