% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cste_indicators.R
\name{cste_indicators}
\alias{cste_indicators}
\title{Add CSTE drug overdose indicators to ICD-10-CM coded hospitalization data}
\usage{
cste_indicators(data, diag_codes=c("DIAG1",  "DIAG2",  "DIAG3",  "DIAG4",  "DIAG5",  "DIAG6",  "DIAG7",  "DIAG8",
                                   "DIAG9",  "DIAG10", "DIAG11", "DIAG12", "DIAG13", "DIAG14", "DIAG15", "DIAG16",
                                   "DIAG17", "DIAG18", "DIAG19", "DIAG20", "DIAG21", "DIAG22", "DIAG23", "DIAG24",
                                   "DIAG25", "ECODE1", "ECODE2", "ECODE3","ECODE4", "ECODE5", "ECODE6"))
}
\arguments{
\item{data}{The dataframe containing claims data.}
\item{diag_codes}{A vector of quoted variable names representing the diagnosis codes in the data. Default is KIPRC's field names.}
}
\description{
This function adds indicator (0,1) variables to emergency department or inpatient hospitalization claims data.
For definitions, see: https://resources.cste.org/ICD-10-CM/Standardized%20Validation%20Datasets/Indicator-Specific%20Regular%20Expressions_2_5_20.pdf
}
\examples{
ip_data %>% cste_indicators()
}
\keyword{ICD-10-CM}
