% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_substances.R
\name{get_substances}
\alias{get_substances}
\title{Extract substance information from free-text information}
\usage{
get_substances(data, ...)
}
\arguments{
\item{data}{The dataframe containing the free text information.}

\item{...}{The free text fields you wish to scan for substances.}
}
\description{
This function creates indicator variables for substances that are listed in free-text. Indicator variables for a
list of drugs and other terms are created in the dataframe supplied to the function.
}
\examples{
get_substances(data=death_certs, lina, lineb, linec, lined, injurydescription, contributingcauses)
}
\keyword{Drugs,}
\keyword{free-text}
\keyword{scan}
\keyword{substances,}
