% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_ods.R
\name{id_ods}
\alias{id_ods}
\title{Subset death certificate records to those with drug overdose death underlying cause of death fields}
\usage{
id_ods(data, ucod_field)
}
\arguments{
\item{data}{The death certificate dataframe.}

\item{ucod_field}{The underlying cause of death field.}
}
\description{
This function subsets a dataframe of death certificates to those with a UCOD field indicating acute drug poisoning.
}
\examples{
id_ods(data=death_certs, field=UCOD)
}
\keyword{Drugs,}
\keyword{cause}
\keyword{certificates}
\keyword{death}
\keyword{death,}
\keyword{of}
\keyword{underlying}
